requirements.txt
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5
Pillow>=9.0
django-import-export>=3.0
django-filter>=22.0
djangorestframework>=3.14
django-widget-tweaks>=1.4
python-dateutil>=2.8
gunicorn>=20.1.0
requests>=2.25.0
pytest>=6.0
pytest-django>=4.0
django-simple-history>=3.5
django-redis>=5.4
psycopg2-binary>=2.9
openpyxl>=3.1
reportlab>=4.4
